CURD operation updated

+++ ServiceXpress_Backend/src/main/java/com/backend/repository/VehicleTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VehicleTypeRepository extends JpaRepository<VehicleType, Integer> {
+public interface VehicleTypeRepository extends JpaRepository<VehicleType, Long> {
 }
